fix: reorg, applied 'use strict', added peer dependencies
