+++ .eslintrc.json
@@ -2,7 +2,6 @@
     "root": true,
     "extends":[
         "eslint-config-moxy/es9",
-        "eslint-config-moxy/addons/es-modules",
         "eslint-config-moxy/addons/node",
         "eslint-config-moxy/addons/jest"
     ]
