+++ lib/playback.js
@@ -1,19 +1,21 @@
+'use strict';
+
 module.exports = (userOptions = {}) => (nextConfig = {}) => (
     {
         ...nextConfig,
         ...{
             webpack(config, options) {
-                const { dev, isServer } = options;
                 const { assetPrefix } = nextConfig;
+                const { dev, isServer } = options;
 
                 config.module.rules.push({
-                    test: /\.(png|jpg|jpeg|gif|webp|ico)$/,
+                    test: /\.(mp3|flac|wav|aac|ogg|oga|mp4|m4a|webm|ogv)$/,
                     loader: 'url-loader',
                     ...userOptions,
                     options: {
                         limit: 0,
-                        name: dev ? 'images/[name].[ext]' : 'images/[name].[hash:15].[ext]',
-                        publicPath: `${assetPrefix}_next/static/chunks/media`,
+                        name: dev ? 'playback/[name].[ext]' : 'playback/[name].[hash:15].ext',
+                        publicPath: `${assetPrefix}/_next/static/chunks/media`,
                         outputPath: 'static/chunks/media',
                         emitFile: !isServer,
                         ...userOptions.options,
@@ -29,4 +31,3 @@ module.exports = (userOptions = {}) => (nextConfig = {}) => (
         },
     }
 );
-
