+++ lib/playback.test.js
@@ -1,4 +1,6 @@
-const { withRasterImages } = require('../../index.js');
+'use strict';
+
+const { withPlayback } = require('../index.js');
 
 const options = {
     dev: true,
@@ -20,19 +22,19 @@ beforeEach(() => {
 });
 
 it('should match snapshot', () => {
-    const arrayOfImagesRules = [];
+    const arrayOfPlaybackRules = [];
 
-    arrayOfImagesRules.push(withRasterImages()().webpack(config, options).module.rules);
+    arrayOfPlaybackRules.push(withPlayback()().webpack(config, options).module.rules);
 
-    expect(arrayOfImagesRules).toMatchSnapshot();
+    expect(arrayOfPlaybackRules).toMatchSnapshot();
 });
 
 it('should return a single rule', () => {
-    const arrayOfImagesRules = [];
+    const arrayOfPlaybackRules = [];
 
-    arrayOfImagesRules.push(withRasterImages()().webpack(config, options).module.rules);
+    arrayOfPlaybackRules.push(withPlayback()().webpack(config, options).module.rules);
 
-    expect(arrayOfImagesRules).toHaveLength(1);
+    expect(arrayOfPlaybackRules).toHaveLength(1);
 });
 
 it('should return a function if next config is a function', () => {
@@ -40,7 +42,7 @@ it('should return a function if next config is a function', () => {
         webpack: mockFunction,
     };
 
-    withRasterImages()(nextConfig).webpack(config, options);
+    withPlayback()(nextConfig).webpack(config, options);
 
     expect(nextConfig.webpack.mock.calls).toHaveLength(1);
 });
