+++ lib/raster-images.js
@@ -1,19 +1,21 @@
+'use strict';
+
 module.exports = (userOptions = {}) => (nextConfig = {}) => (
     {
         ...nextConfig,
         ...{
             webpack(config, options) {
-                const { assetPrefix } = nextConfig;
                 const { dev, isServer } = options;
+                const { assetPrefix } = nextConfig;
 
                 config.module.rules.push({
-                    test: /\.(mp3|flac|wav|aac|ogg|oga|mp4|m4a|webm|ogv)$/,
+                    test: /\.(png|jpg|jpeg|gif|webp|ico)$/,
                     loader: 'url-loader',
                     ...userOptions,
                     options: {
                         limit: 0,
-                        name: dev ? 'playback/[name].[ext]' : 'playback/[name].[hash:15].ext',
-                        publicPath: `${assetPrefix}/_next/static/chunks/media`,
+                        name: dev ? 'images/[name].[ext]' : 'images/[name].[hash:15].[ext]',
+                        publicPath: `${assetPrefix}_next/static/chunks/media`,
                         outputPath: 'static/chunks/media',
                         emitFile: !isServer,
                         ...userOptions.options,
@@ -29,3 +31,4 @@ module.exports = (userOptions = {}) => (nextConfig = {}) => (
         },
     }
 );
+
