+++ lib/raster-images.test.js
@@ -1,4 +1,6 @@
-const { withPlayback } = require('../../index.js');
+'use strict';
+
+const { withRasterImages } = require('../index.js');
 
 const options = {
     dev: true,
@@ -20,19 +22,19 @@ beforeEach(() => {
 });
 
 it('should match snapshot', () => {
-    const arrayOfPlaybackRules = [];
+    const arrayOfImagesRules = [];
 
-    arrayOfPlaybackRules.push(withPlayback()().webpack(config, options).module.rules);
+    arrayOfImagesRules.push(withRasterImages()().webpack(config, options).module.rules);
 
-    expect(arrayOfPlaybackRules).toMatchSnapshot();
+    expect(arrayOfImagesRules).toMatchSnapshot();
 });
 
 it('should return a single rule', () => {
-    const arrayOfPlaybackRules = [];
+    const arrayOfImagesRules = [];
 
-    arrayOfPlaybackRules.push(withPlayback()().webpack(config, options).module.rules);
+    arrayOfImagesRules.push(withRasterImages()().webpack(config, options).module.rules);
 
-    expect(arrayOfPlaybackRules).toHaveLength(1);
+    expect(arrayOfImagesRules).toHaveLength(1);
 });
 
 it('should return a function if next config is a function', () => {
@@ -40,7 +42,7 @@ it('should return a function if next config is a function', () => {
         webpack: mockFunction,
     };
 
-    withPlayback()(nextConfig).webpack(config, options);
+    withRasterImages()(nextConfig).webpack(config, options);
 
     expect(nextConfig.webpack.mock.calls).toHaveLength(1);
 });
