+++ lib/svg.js
@@ -1,3 +1,5 @@
+'use strict';
+
 module.exports = (pluginOptions = {}) => (nextConfig = {}) => (
     {
         ...nextConfig,
