+++ lib/svg.test.js
@@ -1,4 +1,6 @@
-const { withSVG } = require('../../index.js');
+'use strict';
+
+const { withSVG } = require('../index.js');
 
 const options = {
     dev: true,
