+++ package.json
@@ -28,13 +28,13 @@
   "dependencies": {
     "file-loader": "^4.2.0",
     "raw-loader": "^3.1.0",
-    "standard-version": "^7.0.0",
     "svg-css-modules-loader": "^1.6.1",
     "svgo": "^1.3.0",
     "svgo-loader": "^2.2.1",
     "url-loader": "^2.1.0"
   },
   "devDependencies": {
+    "standard-version": "^7.0.0",
     "@commitlint/cli": "^8.1.0",
     "@commitlint/config-conventional": "^8.1.0",
     "eslint": "^6.3.0",
@@ -42,5 +42,8 @@
     "husky": "^3.0.5",
     "jest": "^24.9.0",
     "lint-staged": "^9.2.5"
+  },
+  "peerDependencies": {
+    "next": "^9.0.0"
   }
 }
